Add files via upload
The Simple Arduino/Python IV curve tracer is intended to be used with a prototype BSCCO Intrinsic Josephson Junction (IJJ)  Superconducting Single Photon Detector (SSPD), operated using a 90K cryocooler. By Ohm's Law, V = IR, a change in resistance caused by a single photon interacting with the superconducting current in the IJJ will create a spike in voltage. Moreover, in an IJJ, the tunneling cooper pairs broken by the incoming single photon should itself entangle the electrons and their subsequent microwave pairs. This creates an interesting multiplication of the source entanglement, from say an SPDC light source of entangled photons. If one treats the IJJ element as a diode in the IV curve tracer, one should be able to record the Voltage pulses.